clean up, update docs, remove flatpickr
