+++ package.json
@@ -64,7 +64,6 @@
     "@maskito/kit": "^3.5.0",
     "@maskito/vue": "^3.5.0",
     "axios": "^1.5.0",
-    "flatpickr": "^4.6.9",
     "reka-ui": "^2.9.9"
   },
   "peerDependencies": {
