+++ src/composables/index.ts
@@ -57,6 +57,29 @@ export {
   useBaseAPIDelete,
 }
 
+// dateRange
+import {
+  calendarDateToUnix,
+  dateRangeLast7Days,
+  dateRangeLast30Days,
+  dateRangeThisMonth,
+  dateRangeLastMonth,
+  dateRangeThisYear,
+  dateRangeLastYear,
+  dateRangeClear,
+} from "./dateRange"
+
+export {
+  calendarDateToUnix,
+  dateRangeLast7Days,
+  dateRangeLast30Days,
+  dateRangeThisMonth,
+  dateRangeLastMonth,
+  dateRangeThisYear,
+  dateRangeLastYear,
+  dateRangeClear,
+}
+
 // flashes: Only expose the useFlashes composable for custom flashing and useAppFlasher.
 import { useFlashes, useAppFlashes, useAppFlasher } from "./useFlashes"
 import type { FlashMessage, FlashType, FlashStore, Flasher } from "./useFlashes"
