+++ src/lib-components/index.ts
@@ -34,7 +34,6 @@ import { default as ProgressCirclesLabeled } from "./indicators/ProgressCirclesL
 import { default as BaseInput } from "./forms/BaseInput.vue"
 import { default as Checkbox } from "./forms/Checkbox.vue"
 import { default as DateRangePicker } from "./forms/DateRangePicker.vue"
-import { default as DateRangePickerV2 } from "./forms/DateRangePickerV2.vue"
 import { default as DateTime } from "./forms/DateTimeInput.vue"
 import { default as InputError } from "./forms/InputError.vue"
 import { default as InputHelp } from "./forms/InputHelp.vue"
@@ -78,7 +77,6 @@ export {
   BaseInput,
   Checkbox,
   DateRangePicker,
-  DateRangePickerV2,
   DateTime,
   InputError,
   InputHelp,
@@ -127,7 +125,6 @@ export interface TreesComponents {
   BaseInput: typeof BaseInput
   Checkbox: typeof Checkbox
   DateRangePicker: typeof DateRangePicker
-  DateRangePickerV2: typeof DateRangePickerV2
   DateTime: typeof DateTime
   InputError: typeof InputError
   InputHelp: typeof InputHelp
